REIS-197 additional fix
