+++ app/component/TimetableWeekViewPdf.js
@@ -13,6 +13,11 @@ import {
 import { FormattedMessage } from 'react-intl';
 import moment from 'moment';
 import Icon from './Icon';
+import Relay from "react-relay/classic";
+import take from "lodash/take";
+import {routeNameCompare} from "../util/searchUtils";
+import StopPageHeader from "./StopPageHeader";
+import StopCardHeaderContainer from "./StopCardHeaderContainer";
 
 Font.register({
   family: 'Lato',
@@ -161,7 +166,7 @@ function TextDoubleCell(props) {
   );
 }
 
-function TimetableWeekViewPdf({ patterns }) {
+export default function TimetableWeekViewPdf({ patterns }) {
   let hasDifferentArrivalDepartures = false;
 
   const renderTime = tripTime => {
@@ -405,67 +410,3 @@ function TimetableWeekViewPdf({ patterns }) {
 TimetableWeekViewPdf.propTypes = {
   patterns: PropTypes.arrayOf(PropTypes.object).isRequired,
 };
-
-export default function PDFButton(props) {
-  const [isGenerating, setIsGenerating] = useState(false);
-
-  const renderPDF = () => {
-    setIsGenerating(true);
-
-    setTimeout(() => {
-      const docBlob = pdf(TimetableWeekViewPdf(props)).toBlob();
-
-      docBlob
-        .then(blob => {
-          // create a blobURI pointing to our Blob
-          const blobUrl = URL.createObjectURL(blob);
-          const link = document.createElement('a');
-          link.href = blobUrl;
-          link.onclick = () => {
-            window.open(blobUrl);
-            return false;
-          };
-          // some browser needs the anchor to be in the doc
-          document.body.append(link);
-          link.click();
-          link.remove();
-          // in case the Blob uses a lot of memory
-          // setTimeout(() => URL.revokeObjectURL(blobUrl), 7000);
-        })
-        .catch(err => {
-          // TODO: show error result
-          console.error(err);
-          alert('couldnt generate your pdf, please try again later');
-        })
-        .finally(() => {
-          setIsGenerating(false);
-        });
-    }, 50);
-  };
-
-  return (
-    <button
-      className="secondary-button small"
-      style={{
-        fontSize: '0.8125rem',
-        textDecoration: 'none',
-        marginBottom: '0.7em',
-        marginLeft: 'auto',
-        marginRight: '0.7em',
-        display: 'flex !important',
-      }}
-      disabled={isGenerating}
-      onClick={renderPDF}
-    >
-      <Icon img="icon-icon_print" />
-      {isGenerating ? (
-        <FormattedMessage id="loading" defaultMessage="Loading..." />
-      ) : (
-        <FormattedMessage
-          id="print-timetable-plan"
-          defaultMessage="Koondplaan"
-        />
-      )}
-    </button>
-  );
-}
